Added utility add_rl
